file extension mismatch was causing build errors
The file was looking for .jpg (note, in lowercase) but could not find this as the only file of that name is .JPG (note, in uppercase). This caused the site not to build or reflect the changes that were made. Your changes should be live after I merge this update @RouguiFuta